+++ client/src/data/index.ts
@@ -1,1 +1,2 @@
 // Static content data will be exported from here
+export { personal } from './personal'